mcount: Implement full dynamic tracing on x86_64 (w/ capstone)
when target program did not compiled with compiler support like '-pg'
then use disassembly instead, to determine whether possible to use
instrumentation or not. we can patch the function prolog's instructions
when it can be patched more than 5-bytes.

Full-dynamic consists of the following process.
1. instrument __dentry__ trampoline to end of text segment.
2. Store the instructions which located at the prolog of the function,
if patchable.
3. instrument the 'relative address call instruction' to 'prolog of the
function' to call the trampoline.

In the step 2, append the jump code and address to return to the
original execution flow.

With full-dynamic, the execution flow changes as follows:
1. Calling __dentry__ after a function call
2. Call mcount_entry from __dentry__ (perform a uftrace operation)
3. Return to the stored instructions.
4. Execute the stored instructions and jump to the return address.

Signed-off-by: Hanbum Park <kese111@gmail.com>
Signed-off-by: Namhyung Kim <namhyung@gmail.com>