replay: Make arguments and retvals colorful
Automatic argument display is great but it sometimes looks too verbose
because of too many arguments and return values.

This patch makes arguments and return values more colorful if they are
some special types. The color scheme is as follows:

  - string types     : magenta
  - char types       : magenta
  - enum types       : blue
  - function pointer : cyan
  - parens and comma : bold

This would make users more comfortable to identify those values.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>