misc: Do not read string from dest buffers
There are some char* type arguments, which are useless reading it as
string in some cases.  It's because the contents of its target pointer
is unconditionally being overwritten inside the given function below:

  int sprintf(char *str, const char *format, ...);
  int snprintf(char *str, size_t size, const char *format, ...);
  char *fgets(char *s, int size, FILE *stream);

This patch changes such arguments from char* to void* so that uftrace
doesn't have to dereference and store the string during record.

Suggested-by: Namhyung Kim <namhyung@gmail.com>
Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>