tui: Fix segfault when filter is used
Honggyu reported that tui crashed when filter is used.  It was from
add_remaining_node() called at the end of processing.  If a task has no
function due to filter, fork() exit entry will remain and it makes to
caller the function with no graph setup for the task.

Fixed: #611

Signed-off-by: Namhyung Kim <namhyung@gmail.com>