trigger: Allow to ignore (register) type of argspec
When it reads data from a different machine (cpu architecture), it can
have a different set of registers that host's.  But it makes to fail to
read due to the parsing error.  As it's already recorded we don't care
the type of argspec (which says where it retrieves the data).  The only
think we want to know is the size of data.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>