replay: Implement 'name' trigger
The name trigger is to set a arbitrary name for some value get from
arguments or return value.  By adding name users can identify them
easily from other.  I think it might be helpful to use it for
initializer or allocator of some kind of object.

User can give a name for an object and it'll be used as a prefix if
there are multiple objects.  As it thinks the value as memory address,
it'll overwrite the name with new one when same value is used (guessing
the memory was freed and reused for other purpose).

For a simple example, the following will name the return value of
alloc1() as 'myobj'.  After that as function will use the value
automatically show it with the name.

  $ uftrace -a -R alloc1@name=myobj -D2 tests/t-allocfree
  # DURATION     TID     FUNCTION
              [ 19713] | main(1, 0x7ffcbc01dd98) {
     6.540 us [ 19713] |   alloc1(1) = myobj;
     2.702 us [ 19713] |   free1(myobj);
    15.130 us [ 19713] | } = 0; /* main */

If you want to use it at the function entry, the name spec should be
followed by ':' and arg spec (like arg1).

  $ uftrace -a -A free1@name=myobj:arg1 -D2 tests/t-allocfree
  # DURATION     TID     FUNCTION
              [ 19730] | main(1, 0x7ffeaa673328) {
     7.514 us [ 19730] |   alloc1(1) = 0x5585a66e18f0;
     3.169 us [ 19730] |   free1(myobj);
    17.343 us [ 19730] | } = 0; /* main */

Signed-off-by: Namhyung Kim <namhyung@gmail.com>