replay: Allow overwrite name trigger
When name trigger is used at record time, the argspec (including
namespec) is saved in the data file so it's always used.  If user want
to apply other name, it should have a way to override it on replay time.

So allow -A and -R options on replay but only limit it to name specs.
Otherwise it could confuse the fstack reading arguments and return value
with wrong size or type info.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>