arch/x86: Fix segfault in test 161 (pltbind_now)
When -z now and -z reloc linker options are used for the ET_EXEC, it'll
access invalid symbol address in mcount_arch_hook_no_plt().  It's
because the address of symbol wasn't adjusted for the (load) offset.
But for ET_EXEC dl_iterate_phdr() returns 0 (offset).  I think it needs
to use sh_addr instead of sh_off.  For ET_DYN they should be same.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>