uftrace: Add include/uftrace.h for variable recording
Recording funtcion arguments and return values is one of the core
features of uftrace and it provides useful information to users.

In some cases, however, user may want to record values of some specific
variables to better understand the execution flow.  It's especially
useful when recording a member variable inside struct and class types.

This patch introduces a header file "include/uftrace.h" so that users
can include it and record some values in the middle of functions.

The usage is as follows:

  1. Include "uftrace.h" in the target source code.
  2. Modify the source code by inserting uftrace_print() function with a
     single argument that you want to record with its function flow.
  3. Run "uftrace record" for the target program.
  4. Run "uftrace replay" to observe the values.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>