record: write the previous memory area to file when reach special mapped area.
it will ignore the previous map area if we continue when reach special
mapped area like [stack]. because typically the [stack] area that
specially mapped is end of memory space. therefore, modify codes of
record_proc_maps function to save the previous memory area to file when
reach special area.

Signed-off-by: Hanbum Park <kese111@gmail.com>