fstack: Handle perf sched-in event before kernel function
When deep kernel tracing is enabled, perf sched events will be
intermixed with kernel functions.  Problem is that the sched-out/in pair
come with a different stack depth so it confuses fstack accounting.

Ideally, the sched-in event should come before any kernel function so
when it sees a kernel function after sched-out (and before sched-in)
peek the matching sched-in event and handle it before the kernel
function.  It'll make processing side like report and graph easier.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>