tui: Open file using ctags if debug info isn't available
Sometimes it may not be easy to have debug info from the recorded data.
It'd be useful to provide another way to open editor using ctags.

It's especially useful when opening kernel functions because there's no
way to handle them as of now.

The 'O' key first tries to open file based on debug info, then tries to
use ctags.  If both of them are failed, then does nothing.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>