tui: Fix line wrapping problem due to box chars
Currently, tui calculates its width based on COLS, but it doesn't
consider the size of box characters.  A box character spends 3 bytes
but it's displayed as a single character, so the previously width
calculation doesn't match the actual width printed on screen.

The approach of this patch is to print the characters on a global buffer
"linebuf" first, then cut the end of line based on COLS after
considering the width of box characters.  Then the "linebuf" is printed
in a single line without wrapping around to the next line unexpectedly.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>