symbol: Assign correct address for dynamic symbols
In case of a dynamic library, it can call functions defined in itself
through PLT.  In that case, a symbol assigned to a PLT will have an
address of defined function in the text section.

But on ARM (I guess it's THUMB2) PLT entries in an executable don't have
fixed size and some (undefined) dynamic symbols have an address for PLT
location.  Also ELF spec says that it's ok for an undefined symbol to
have a value of address of the PLT (when it needs an address of an
external library function).

This makes confusion when loading dynamic symbols and resulted in an
incorrectly mixed symbol table after merging.  Use symbol's value as an
address only if it's an undefined symbol.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>