symbol: Save a dummy symbol for end of the functions
As we save function symbols and data symbols together in a single
symbol table.  It might be confusing that last function has a big size
when replay.  This also led to a wrong symbol resolution in some
cases.  Add a marker for the end of function to find such cases.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>