uftrace: Revert 'shell-support' branch
It seems to be tricky to handle shell script tracing.  Especially, the
problem looks clear when tracing a big program such as chrome.

It shows too many small ignorable short trace records and it increases
the number of tasks too much.

To avoid such problems, this patch reverts the following commits:

  0eab415d Merge branch 'shell-support'
   85c3202f test: Add a test case for script tracing
   8f910e8b record: Support tracing script files
   ae3faff7 symbol: Add check_script_file()

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>