info: Use --task option to show task tree
Honggyu requested an option to show task tree so that it can
distinguish interesting tasks easily.  Add list nodes in the
uftrace_task struct to build tree hierachy and first_task
pointer into the session link struct.

The new task info will show the first timestamp of the task (usually
when the task was created), two flags that indicate whether the task
was forked (instead of clone thread) and it has session (meaning that
it called exec* syscall).  And basic task info like tid and name, and
finally the actual data size belongs to the task.  This is a quick
operation doesn't need to read all trace data but the metadata.

For example simple parent-child tasks will be displayed like below:

  $ uftrace info --task
  #      TIMESTAMP       FLAGS     TID    TASK      DATA SIZE
       399486.071314857  FS     [ 91331]  parent        0.012 MB
       399486.077029725  F      [ 91333]  child         0.016 MB

Signed-off-by: Namhyung Kim <namhyung@gmail.com>