symbol: Pre-allocate symbol table before loading
When we load symbol tables from an ELF binary, it can know the upper
bound of symbol table size.  So instead of increasing the table
gradually, preallocate the table once and then resizing after full
table is loaded.  This way, it can reduce number of calls to realloc()
which needs to copy existing memory block everytime.

Fixed: #723

Signed-off-by: Namhyung Kim <namhyung@gmail.com>