arch/x86_64: restores all registers after calling dentry
previously, only some registers were restored when calling dentry
according to the calling convence. however, if an option such as '-pg'
or '-finstrument-function' is not specified, compiler uses registers
that are not used by the called function.

if do not restore the register in this situation, it has the same effect
as changing the register used as a variable in the callee function. in
other words, unintentionally changing variables in the callee function.

Fixed: #731

Signed-off-by: Hanbum Park <kese111@gmail.com>