arch/x86_64: Fix crash on (full) dynamic tracing
Hanbum found that program could crash when tracing binary due to
compiler optimization.  With -fipa-ra option gcc can generate code to
access caller-saved registers across function call if the register is
not changed inside the (child) function.  This breaks assumption that
uftrace holds and resulted in crashes.

Note that this problem is not exist when built with -pg since it adds
an external call (mcount) which compiler cannot see.

Fixed: #731

Analyzed-by: Hanbum Park <kese111@gmail.com>
Signed-off-by: Namhyung Kim <namhyung@gmail.com>