chrome: Fix a bug when displaying character type
It didn't fully handle the char type dump when --chrome option is used.
The following dump output gets failed when loading it in chrome browser.

  {...,"args":{"arguments":"('"')"}},

It's because the second argument '"' is misinterpreted as " is the
content, but it looks like a closing quotation mark.  The " character
should be escaped from '"' to '\"' as follows:

  {...,"args":{"arguments":"('\"')"}},

This also handles \ character to be escaped because the following json
cannot be interpreted in chrome browser.

  {...,"args":{"arguments":"(\"Next\:\")"}},

It becomes as follows:

  {...,"args":{"arguments":"(\"Next\\:\")"}},

Fixed: #735

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>