record: Use absolute path of uftrace.data
Current working directory can be changed by calling chdir.
If it is changed, the problem that uftrace.data can't be found can occur:

  $ uftrace record --force ./configure
  mcount: /home/honggyu/work/uftrace/libmcount/record.c:1093:record_proc_maps
   ERROR: cannot open for writing maps file: No such file or directory

Because of trying to create "uftrace.data/sid-xxxxx.map" of which the prefix
is a relative path. So convert uftrace data directory path that can be
relative to an absolute path for both scripts and target programs that
can call chdir.

Fixed: #745

Reported-by: Honggyu Kim <honggyu.kp@gmail.com>
Signed-off-by: Taeung Song <treeze.taeung@gmail.com>