plthook: Fix infinite loop in dlsym from shared library
Honggyu reported that there's an infinite loop after dlsym() when it
's called with RTLD_NEXT from a shared library.  The RTLD_NEXT
searches the symbol after the calling module in a library loading
order.  It knows the caller by looking at the return address of the
dlsym() function.

But as plthook codes replaces the return addrese to plt_hooker (in
libmcount), the dynamic linker thinks that it's called from the
libmcount instead of the wrapping library and returns the address of
the library.  This make the infinite loop since usually the wrapping
library calls the real function using the address returned by
dlsym().

Since it's not the common case, add PLT_FL_DLSYM flag and handle the
problematic case by recovering the return address.  As we want to
trace the dlsym() after the case, resolve address of dlsym() first so
that the PLT entry won't be overwritten by dynamic linker.

Fixed: #714

Signed-off-by: Namhyung Kim <namhyung@gmail.com>