dynamic: support libraries for dynamic tracing
The dynamic patching was applicable to main executable only.
However it can also patch the functions in the shared objects.

Usage:

$ uftrace -P getpid@libc-2.27.so t-abc

Name of function to be patched comes before @ and the name
of the module containing the function follows. Above example
enables getpid function in the libc-2.27.so.

[notice]
library name must be mapped library name that mean actually library
name. it could not work with alias name.

Signed-off-by: Hanbum Park <kese111@gmail.com>