uftrace: Use 'default.opts' to check whether it's uftrace directory
It used to check 'info' file in recorded directory to decided if it's
uftrace directory or not.

However, in many cases, some broken data directory only stored uftrace
recorded data but without 'info' file due to some crashes.

It'd be better to check 'default.opts' as well because it's also created
right after uftrace.data is created.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>