aarch64: Handle struct return type by keeping more regs on stack
Current implementation for mcount_return and plthook_return only reserve
x0 register assuming the size of return value is enough with 8 bytes.

But some functions may have a return type in struct and its size is more
than 8 bytes as following example:

  struct { int64_t quot, int64_t rem}
         __aeabi_ldivmod(int64_t numerator, int64_t denominator) {
    int64_t rem, quot;
    quot = __divmoddi4(numerator, denominator, &rem);
    return {quot, rem};
  }

In this case, we might partially lose return values because the return
value is stored in x0 and x1 registers to keep 16 bytes in the return
type of struct { int64_t quot, int64_t rem}.

So this patch fixes to keep both x0 and x1 registers to prevent such
problems.

It's also required to save d0 and d1 VFP registers in return hook.

This patch is inspired by the following commit for arm.

  ffb69ce arm: Handle struct return type by keeping more regs on stack

Fixed: #791

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>