build: fix a build error for i386
There is a format string in libcapstone.c that did not match the type.
because of it, configuration was not done as intended. to fix this
issue, current format string "%lu" was changed to "%zu".

function 'mcount_patch_func' declaration was changed after size filter
updated. so, to solve conflict its declaration inside of i386 arch also
changed.

Fixed : #805

Signed-off-by: Hanbum Park <kese111@gmail.com>