symbol: fix bug in unittest for i386
it expect that "info->dlpi_addr" have the virtual address of the main
binary. but it was 0 in i386. so, added some code for guess the virtual
address of main binary.

main binary symbol has the absolute address not the relative address in
i386. so, absolute address of main binary symbol was compared with
relative address of function "add_map" at testcase symbol_load_map.

like belows:
[waitpid start] 804b150 < [add_map] 604f4 < [waitpid end] 804b160
[waitpid start] 804b150 < [add_map] 604f4 < [waitpid end] 804b160
[strsep start] 804b140 < [add_map] 604f4 < [strsep end] 804b150
[strsep start] 804b140 < [add_map] 604f4 < [strsep end] 804b150

to solve this issue, added SYMTAB_FL_ADJ_OFFSET to symtabs flag.

Signed-off-by: Hanbum Park <kese111@gmail.com>