dynamic: Fix error when used with size filter
Honggyu reported that nodejs failed to run when some functions are
filtered by -Z option.  The mcount_find_code() was used originally in
the __dentry__ asm function and it should find an address to jump for
the copied instructions.  So it used pr_err() when it didn't find an
address in the code tree.

However the size filter code also calls mcount_find_code() now and
it's possible there's no entry for the address.  So change the code to
return NULL pointer and get rid of the pr_err() inside.  If __dentry__
has the NULL pointer (which should not happen anyway) we can identify
the problem since RIP would be 0.

Fixed: #822

Signed-off-by: Namhyung Kim <namhyung@gmail.com>