utils: Print crash messages with pr_warn
In some cases, it's difficult to analyze a bug report based on logfile
output because it doesn't contain some critical messages, which are
printed by pr_color.  In most cases, such colored messages are crash
points, but it's difficult to match where the program gets segfault
among the long list of debug messages.

This patch changes colored messages to pr_warn so that it can be printed
in the logfile if -v is given.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>