record: Remove a message for installtion path check
In some cases, the following message is printed in a normal execution.

  # uftrace record --force pwd
  WARN: Didn't you run 'make install' ?
  /home/root

The message is printed if libmcount.so is not found the installtion
path, but it can be used anywhere especially in embedded environments
that use read-only file system.

It seems that specifying the basename of libmcount.so is enough for
LD_PRELOAD, it'd be better to remove this message.

But if libmcount.so is found in the installtion path, it has the higher
priority of use.

Fixed: #796

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>