dynamic: Disallow patching functions with an indirect jump
Since it's hard to predict the target of an indirect jump, we just disallow patching it for now.

"Kprobes analyzes the entire function and verifies that there is no
jump into the optimized region. Specifically, the function contains no
indirect jump." See section 'Safety Check' in kprobe documentation for more details [1].
This patch reduces the success rate by approximately 3% when tracing the executable of uftrace.

[1] https://github.com/torvalds/linux/blob/master/Documentation/kprobes.txt

Signed-off-by: Anas Balboul <anasbalbo@gmail.com>