mcount: Allow full-dynamic tracing to instrument unsupported functions on x86_64 (w/ capstone)
Some functions can't be instrumented by full-dynamic tracing, for safety
reasons. Functions that jump to a prologue that has been modified may
cause an undefined behavior. To be able to instrument these functions, one
can embed an illegal instruction (for example "int3") in the head of every
instruction that has been moved. This way, if a thread branch to the function
prologue, it will step on the illegal instruction and a handler will be called
by the kernel to redirect the thread to the original instruction. Afterwards,
the thread will jump back to the function and resume its execution normally.

Instrumenting unsupported functions process is similar to full-dynamic
tracing process excepts some differences:
1. Create a constraint based on the position of the "int3" in
the offset of the call instruction.
2. Store the instructions located at the prolog of the function,
if patchable.
3. Find and allocate a free address that respect the constraint
created previously.
4. Store the position of "int3" and the address of the original
instruction.
5. instrument the 'relative address call instruction' to 'prolog of the
function' to call the trampoline.

The execution flow is similar to the full-dynamic tracing except when
a thread branch to a prologue and step on an "int3":

1. The trap handler is called.
2. The address from where the trap handler was raised is computed.
3. If the trap handler was raised from an address related to our tracepoint,
the thread is redirected to the original instruction.
Else, the original handler (set by the user) is called and the execution is
resumed.

Signed-off-by: Anas Balboul <anasbalbo@gmail.com>