mcount: fix segfault in mcount_get_stack_arg()
If the given argument index is out of memory or inaccessible, a segfault will occur.
I added code to check the validity of the memory address before accessing the address.
Changed check_mem_region() to a global function.
If the index of the argument is invalid, ctx->val.v is padded with zeros.

Fixed: #858

Signed-off-by: MinJeong Kim <98nba@naver.com>