arch/aarch64: Handle PC-relative addressing in ADRP insn
The ADRP instruction is often used at prologue to load an arbitrary
memory address from PIC code.  Converting this to LDR from adjacent
memory which has the calculated the target address.

Note that it needs to have data byte in the middle of saved code.  So
we cannot just put the data right after the LDR instruction since it
would be executed.  Therefore we put it after the jump instruction at
the end of the saved code (which already includes 8 data bytes).  Thus
the final offset would be 24 (bytes).

As have two modified instructions in a prologue would make the problem
more complex, it just reject ADR[P] instructions if it's not the first
instruction of the function.

This increased the success rate by 10% for C-python interpreter.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>