utils: Implement luajit script engine
luajit(https://luajit.org/) is a JIT supported version of lua5.1

Although this is not a standard distribution of lua,
However it is the most widely used version because it is JIT supported.
And you can also use luajit's built-in ffi api to use c code.

It is used in many projects supported by big companies.
following is the list of projects or companies using luajit:
  openresty(https://github.com/openresty)
  cloudflare(https://blog.cloudflare.com/more-consistent-luajit-performance)
  torch(http://torch.ch)

With this adoption of luajit, you can run uftrace scripts faster.
This will be especially useful when you need to write some complicated scripts

Signed-off-by: Byeonggon Lee <gonny952@gmail.com>