uftrace: Fix fread() error handling
The following commit is a regression that breaks every normal recording.

  5cffda2 uftrace: Fix fread() error handling

Unfortunately, this change breaks normal recording as follows:

  $ uftrace t-abc
  uftrace: /home/honggyu/work/uftrace/utils/data-file.c:493:open_info_file
   ERROR: cannot read uftrace header info!

In uftrace.data/info file, cmdline line is not found at all.

The man page of fread describes as follows:

  fread() does not distinguish between end-of-file and error, and callers
  must use feof(3) and ferror(3) to determine which occurred.

In cmds/info.c checking ret != sizeof(fha->buf) doesn't always mean that
an error occured.  If the buffer is large enough, return value of fread()
can be smaller than the size of the buffer.

In utils/symbol.c when while loop ends, len will be always zero.
ferror() can be used to check if an error actually occured in this case.

There are compile warnings as well.

  cmds/info.c: In function ‘fill_cmdline’:
  cmds/info.c:230:25: warning: comparison of integer expressions of different signedness:
                      ‘int’ and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]
    230 |  for (i = 0, p = buf; i < ret; i++, p++) {
        |                         ^
  cmds/info.c:237:22: warning: passing argument 2 of ‘json_quote’ from
                      incompatible pointer type [-Wincompatible-pointer-types]
    237 |  p = json_quote(buf, &ret);
        |                      ^~~~
        |                      |
        |                      size_t * {aka long unsigned int *}
  In file included from /home/gonapps/workspace/uftrace/utils/symbol.h:15,
                   from /home/gonapps/workspace/uftrace/uftrace.h:11,
                   from /home/gonapps/workspace/uftrace/cmds/info.c:20:
  utils/utils.h:337:8: note: expected ‘int *’ but argument is of type ‘size_t *’
                       {aka ‘long unsigned int *’}
    337 | char * json_quote(char *str, int *len);
        |        ^~~~~~~~~~
  cmds/info.c:241:30: warning: comparison of integer expressions of different signedness:
                      ‘ssize_t’ {aka ‘long int’} and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]

The warnings can simply be suppressed by changing ret variable to int type.

Fixed: #952

Signed-off-by: Byeonggon Lee <gonny952@gmail.com>