arch/x86_64: Handle CALL instruction in dynamic patch
The CALL instructions is basically same as PUSH + JMP.  As we can add
JMP at the end already, convert CALL into the two instructions.  The
pushed return address can be calculated from instruction address +
size, and we saved it just after the JMP instruction (offset = 6).

And the JMP instruction also needs to be modified due to the offset
change.  So ignore the JMP in patch_normal_func() using info->has_jump
field.  This increased ~1% of success rate of the C-python interpreter.

Closed: #988

Signed-off-by: Namhyung Kim <namhyung@gmail.com>