dynamic: Relocate relative jumps
Relative conditional branches jcc8 jumps to a trampoline if condition
is met. The trampoline is an absolute indirect jump that can reach the
target of the original instruction. Relative conditional branches jcc32
are replaced with its equivalent jcc8 before being relocated just like
jcc8. Relocation of relative jumps is achieved by directly replacing it
with an absolute indirect jump to the target of the original
instruction.

Signed-off-by: Anas Balboul <anasbalbo@gmail.com>