dynamic: refactoring do_dynamic_update to support tracing for dynamic loaded module
the variable named 'patterns' maintains a list of module and function
names to be instrumented. previous uftrace do instrumentation only once
when the program started. However, uftrace now plans to support tracing
for dynamic loaded modules. so, need to keep the 'patterns'. for this,
the 'pattern' scope has been changed to global.

at the start of the program, instrumentation was performed in the
'do_dynamic_update' function. The 'do_dynamic_update' function also
creates the 'patterns'. in order to support instrumentation for dynamic
load modules, the 'instrumentation logic' has been split into function
'patch_func_matched'.

Signed-off-by: Hanbum Park <kese111@gmail.com>