dynamic: enable re-allocating for code_page
Previously uftrace runs the dynamic instrumentation once when the
process starts.  And uftrace does not expect to reallocate from
code_page, so it removes write permission from code_page.

However, we need to support dynamic tracing for dynamically loaded
modules (by dlopen).  Changing the permission of the existing of pages
are dangerous so add a new code page for them and as ->frozen field to
indicate pages already processed.

Signed-off-by: Hanbum Park <kese111@gmail.com>
Signed-off-by: Namhyung Kim <namhyung@gmail.com>