fstack: Fix replay with read trigger
When it read tracing data from the files, it saves records for each
task's rstack_list to handle time filters rather than printing it out
immediately.  With this approach, we can skip functions under time
filter when replay.

The moment it read an EXIT record it can compare function execution
time with the filter to decide whether it prints the function or not.
This requires a matching ENTRY record to calculate execution time.  If
it's not found, that means the current function is already handled by
a previous (sibling) function so it's ok to print it.

But there was a hole in this logic when the rstack_list only has EVENT
records.  In that case, it was stuck in an infinite loop finding an
ENTRY record in the list.  We should traverse the list (in reverse
order) once to find it and bail out if there's no matching record.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>