build: Find and use a correct python3 version
It might be failed to find a correct python3 version used for scripting.
It's because python3 version is found by the python3 binary, but not by
finding python3 library actually used.

The current implementation finds python3 version as follows:

  $ python3 -c 'import sysconfig; print(sysconfig.get_config_vars()["LDVERSION"])'
  3.7m

However, python3-config finds different version.

This patch is to find the correct version by parsing the output of
'python3-config --libs'.

  $ python3-config --libs | sed 's/-l\(python[0-9\.m]\+\) .*/\1/g'
  python3.5m

  $ python3.7-config --libs | sed 's/-l\(python[0-9\.m]\+\) .*/\1/g'
  python3.7m

This might be a corner case fix but to support many more different
environments.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>