script: Show an warning if interpreter is not available
The current implementation doesn't show anything if scripting
interpreter is not available and it's silently failed.

  $ ./configure --without-libpython
  $ make install
  $ uftrace record t-abc
  $ uftrace script -S scripts/count.py
  (... print nothing ...)

It just prints the message only when verbose option is given, but most
of the new users may not know what goes wrong in this way.

  $ uftrace script -S scripts/count.py --debug-domain script:1
  script: failed to init python scripting

Instead of hiding this failure message, this patch changes it from
pr_dbg to pr_warn so that anyone can easily see the problem.

  $ uftrace script -S scripts/count.py --debug-domain script:1
  WARN: failed to init python scripting

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>