tests: Fix script engine tests
The current tests incorrectly written in its 'pre' function.
It only tests it by running 'uftrace script' and checks whether it
prints 'WARN' message or not, but the actual output is different.

  $ uftrace script
  Usage: uftrace script (-S|--script) <script_file>

It doesn't show WARN message even if the script engine is not available
and the tests are always failed instead of skipped.

This patch fixes the 'pre' function so that it can correctly check if
the script engine is available.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>