doc: Fix configure with env variables in INSTALL.md
The current configure script is tricky to use environmental variables
with its builtin options such as --without-libpython.

As written in INSTALL.md, configure can be used with ARCH and CFLAGS,
but the order is important when used with configure options.

The current direction is written as follows:

  $ ./configure ARCH=arm CFLAGS='--sysroot /path/to/sysroot'

It does work, but it's broken whtn it's used with configure option as
follows:

  $ ./configure ARCH=arm CFLAGS='--sysroot /path/to/sysroot' --without-libpython
  ./configure: line 90: eval: --: invalid option
  eval: usage: eval [arg ...]
  uftrace detected system features:
  ...         prefix: /usr/local
  ...         libelf: [ on  ] - more flexible ELF data handling

So this patch fixes the usage by changing the order of environmental
variables as follows:

  $ ARCH=arm CFLAGS='--sysroot /path/to/sysroot' ./configure

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>