utils: Detect busybox and disable color
In some cases, busybox is used with a script file.

The example case is as follows:

  $ realpath /usr/bin/less
  /usr/lib64/busybox/usr/bin/less

However, the file is not an ELF executable, but just a script.

  $ cat /usr/lib64/busybox/usr/bin/less
  #!/bin/busybox.nosuid

To detect such case, uftrace has to check whether the file is ELF
executable or not.  If it's just a script, then disable color.

Fixed: #1123

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>