utils: Lower debug level in demangler and symbol
In most cases, demangler error is not a critical issue, but it looks
very annoying to see a lot of demangler failures in debug level 1.

In addition, symbol info can also be printed in a higher debug level.

This patch lowers the debug levels to hide messages which are not
related to tracing execution such as mcount and plthook.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>