fstack: Add fstack_check_filter_done()
The fstack_check_filter() calls fstack_{enter,update,exit} functions
but they need to be called carefully.  The fstack_enter/exit are to
adjust filter conditions and they should be called for each
functions (especially for EXIT even after filter applied). But
fstack_update() is to adjust display depth which is meaningful only if
it's not filtered (and displayed somehow).

The fstack_entry() is supposed to be called *before* an ENTRY record
is processed and fstack_exit() is supposed to be called *after* an
EXIT record is processed.  But the fstack_update() is supposed to be
called *after* an ENTRY record and *before* an EXIT record is
processed.  But the fstack_check_filter() ignored that for EXIT
entries.

One more subtlety was fstack_update() also has some special case stack
count reset code which is for exec*() and setjmp/longjmp() calls.
Ideally they should be moved to a location the stack count
manipulation is happened.  For now, sepeate calls to fstack_update()
into fstack_filter_check_done() as a short term solution.

Fixed: #1128

Signed-off-by: Namhyung Kim <namhyung@gmail.com>