fstack: Delay estimated return after schedule event
As libmcount rstack is not aware of perf events, an estimated return
can have inversed time if there's a schedule event.  It's not possible
to adjust the timestamp from libmcount, so let's do it in the fstack.

When it sees a sched-out event, it'll mark next user record to delay
it after the matching sched-in event.  Once we get the sched-in event,
recalcuate return time with the timestamp in the sched event and the
estimated next functions's timestamp (using current + previous records).

Currently it doesn't care about the kernel functions as they'll have
precise return records.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>