filter: Add -H/--hide option to hide functions without subtree
The -H/--hide option is used to apply notrace filter to some functions without
affecting their subtree functions.  In other words, it simply hides
single depth of function and continue to trace child functions.

This option works only for replay time and 'hide' can be used as a
trigger just like other triggers such as filter.

  $ gcc -pg -o t-abc tests/s-abc.c
  $ uftrace -H main -T c@hide t-abc
  # DURATION     TID     FUNCTION
              [ 13037] | a() {
              [ 13037] |   b() {
     2.600 us [ 13037] |     getpid();
     8.500 us [ 13037] |   } /* b */
    11.400 us [ 13037] | } /* a */

This can be useful when user functions are called from a library, but
don't want to trace functions in the library as the following example.

  $ cat call_once.cc
  #include <mutex>
  std::once_flag flag;
  void foo() {}
  int main() {
      std::call_once(flag, foo);
  }

  $ g++ -pg -O2 -pthread call_once.cc
  $ uftrace record a.out
  $ uftrace replay
  # DURATION     TID     FUNCTION
              [ 19661] | main() {
              [ 19661] |   pthread_once() {
              [ 19661] |     std::call_once::$_1::_FUN() {
     1.900 us [ 19661] |       foo();
     6.700 us [ 19661] |     } /* std::call_once::$_1::_FUN */
    50.900 us [ 19661] |   } /* pthread_once */
    56.100 us [ 19661] | } /* main */

We can simply hide functions in C++ std namespace easily as follows.

  $ uftrace replay -H ^std:: a.out
  # DURATION     TID     FUNCTION
              [ 19666] | main() {
              [ 19666] |   pthread_once() {
     2.300 us [ 19666] |     foo();
    31.400 us [ 19666] |   } /* pthread_once */
    38.000 us [ 19666] | } /* main */

It also supports 'hide' trigger for the same purpose.

Closes: #1169

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>
Signed-off-by: Namhyung Kim <namhyung@gmail.com>