tests: Fix t051_return.py test in aarch64
Here is the t051_return.py test result in aarch64.

  $ ./runtest.py 051
  Start 1 tests with 1 worker
  Test case                 pg             finstrument-fu
  ------------------------: O0 O1 O2 O3 Os O0 O1 O2 O3 Os
  051 return              : NZ NZ NZ NZ NZ NG NG NG NG NG

The -pg tests are failed and they return 1 instead of 0.

It's because q0 register is clobbered by mcount_exit routine.  The
current implementation thinks that storing and loading d0 and d1 does
the same job with storing and loading q0 register itself, but the test
result doesn't actually keep the same value in q0 before and after
mcount_exit() execution.

The simple reproducible program can be found in the github issue below.

Fixed: #1153

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>