dynamic: Determine hashmap size using symbol tables
Honggyu reported that chrome binary crashes during initialization when
dynamic tracing is enabled.  It was due to the free() during expansion
of the hash map but I suspect it's tcmalloc's fault.  Honggyu
confirmed that changing the malloc library removed the problem.

However, we'd better avoid it anyway in the default setting.  And we
can know how much entries it'll have since we know the size of symbol
tables.  So use the value for initial capacity of the hashmap instead
of the hard-coded value.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>