symbol: Add check_symbol_file()
The check_symbol_file() function is to check metadata in the symbol
file header.  It includes full path name and optional build-id.  Note
that usually the build-id is SHA-1 hash which has 40 bytes in string
form but other method (like MD-5) can be used later.

For now we only read at most 40 bytes for the build-id.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>