tests: Fix the test 246 report_srcline2
The test 246 fails because 'lib_a' is captured twice, but one of them is
captured by a library call, but summarized into the same function result
in report output.  So the order can be different for each execution
depending on the time spent on the library call, so let's just ignore
library call for the test.

  t246_report_srcline2: diff result of -pg -O0
  --- expect      2020-09-13 15:08:42.112217325 +0100
  +++ result      2020-09-13 15:08:42.112217325 +0100
  @@ -1,4 +1,4 @@
  -2 lib_a
   1 main s-libmain.c:16
   1 foo s-libmain.c:11
  +2 lib_a
   1 lib_b s-lib.c:15

  246 report_srcline2     : NG

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>